# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "OpenVPN Access Server Management API with Bulk Operations, Advanced Search, and Comprehensive Redis Caching",
        "title": "GoVPN API Enhanced with Advanced Redis Caching",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.2.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/openvpn/bulk/users/disconnect": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Disconnect multiple users from VPN with business logic validation (user exists and is connected)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bulk Operations"
                ],
                "summary": "Bulk disconnect multiple VPN users",
                "parameters": [
                    {
                        "description": "Bulk disconnect users request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.BulkDisconnectUsersRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Users disconnected successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.DisconnectResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request - validation error or no valid users",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid or missing authentication",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error - failed to disconnect users",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/openvpn/config/network": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get comprehensive network configuration including client networks, VPN daemon settings, routing, and NAT configuration",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Configuration"
                ],
                "summary": "Get network configuration",
                "responses": {
                    "200": {
                        "description": "Network configuration retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.NetworkConfigResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid or missing authentication",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error - failed to retrieve network config",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/openvpn/config/server/info": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get basic server information including node type, ports, and cluster configuration",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Configuration"
                ],
                "summary": "Get server information",
                "responses": {
                    "200": {
                        "description": "Server information retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ServerInfoResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid or missing authentication",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error - failed to retrieve server info",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/openvpn/users/{username}/disconnect": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Disconnect a specific user from VPN with business logic validation (user exists and is connected)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Disconnect a single VPN user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username to disconnect",
                        "name": "username",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Disconnect user request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.DisconnectUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User disconnected successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.DisconnectResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request - user not found or not connected",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid or missing authentication",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User not found in system",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error - failed to disconnect user",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.BulkDisconnectUsersRequest": {
            "type": "object",
            "required": [
                "usernames"
            ],
            "properties": {
                "message": {
                    "type": "string",
                    "maxLength": 200,
                    "example": "Maintenance disconnect"
                },
                "usernames": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[\"testuser1\"",
                        " \"testuser2\"]"
                    ]
                }
            }
        },
...
        "dto.DisconnectResponse": {
            "type": "object",
            "properties": {
                "connection_info": {
                    "$ref": "#/definitions/dto.UserConnectionInfo"
                },
                "count": {
                    "type": "integer",
                    "example": 2
                },
                "disconnected_users": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[\"testuser1\"",
                        " \"testuser2\"]"
                    ]
                },
                "message": {
                    "type": "string",
                    "example": "Users disconnected successfully"
                },
                "skipped_users": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[\"offline_user\"]"
                    ]
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "total_requested": {
                    "type": "integer",
                    "example": 3
                },
                "validation_errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.UserValidationError"
                    }
                }
            }
        },
        "dto.DisconnectUserRequest": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "maxLength": 200,
                    "example": "Session terminated by administrator"
                }
            }
        },
...
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "string",
                            "example": "BAD_REQUEST"
                        },
                        "details": {
                            "type": "string",
                            "example": "validation failed"
                        },
                        "message": {
                            "type": "string",
                            "example": "Invalid request format"
                        },
                        "status": {
                            "type": "integer",
                            "example": 400
                        }
                    }
                }
            }
        },
...
        "dto.NetworkConfigResponse": {
            "type": "object",
            "properties": {
                "allow_private_nets6_to_clients": {
                    "type": "boolean",
                    "example": true
                },
                "allow_private_nets_to_clients": {
                    "description": "Advanced Network Settings",
                    "type": "boolean",
                    "example": true
                },
                "client_netmask_bits": {
                    "type": "string",
                    "example": "20"
                },
                "client_network": {
                    "description": "Client Network Settings",
                    "type": "string",
                    "example": "172.27.224.0"
                },
                "group_pool": {
                    "type": "string",
                    "example": "172.27.240.0/20"
                },
                "inter_client": {
                    "type": "boolean",
                    "example": false
                },
                "listen_ip": {
                    "type": "string",
                    "example": "all"
                },
                "message": {
                    "type": "string",
                    "example": "Network config retrieved successfully"
                },
                "mss_fix": {
                    "type": "string",
                    "example": "1350"
                },
                "mtu": {
                    "description": "Network Performance",
                    "type": "string",
                    "example": "1420"
                },
                "nat6_enabled": {
                    "type": "boolean",
                    "example": true
                },
                "nat6_masquerade": {
                    "type": "boolean",
                    "example": false
                },
                "nat_enabled": {
                    "description": "NAT Settings",
                    "type": "boolean",
                    "example": true
                },
                "nat_masquerade": {
                    "type": "boolean",
                    "example": false
                },
                "osi_layer": {
                    "type": "string",
                    "example": "3"
                },
                "private_access": {
                    "type": "string",
                    "example": "no"
                },
                "protocol": {
                    "type": "string",
                    "example": "tcp"
                },
                "reroute_dns": {
                    "type": "boolean",
                    "example": true
                },
                "reroute_gateway": {
                    "description": "Routing Settings",
                    "type": "boolean",
                    "example": true
                },
                "server_ip": {
                    "type": "string",
                    "example": "all"
                },
                "status": {
                    "description": "Status",
                    "type": "string",
                    "example": "optimal"
                },
                "tcp_port": {
                    "description": "VPN Daemon Settings",
                    "type": "string",
                    "example": "443"
                },
                "udp_port": {
                    "type": "string",
                    "example": "1194"
                }
            }
        },
...
        "dto.ServerInfoResponse": {
            "type": "object",
            "properties": {
                "admin_ip_address": {
                    "type": "string",
                    "example": "all"
                },
                "admin_port": {
                    "type": "string",
                    "example": "943"
                },
                "client_ip_address": {
                    "type": "string",
                    "example": "all"
                },
                "client_port": {
                    "type": "string",
                    "example": "943"
                },
                "cluster_mode": {
                    "type": "string",
                    "example": "False"
                },
                "failover_mode": {
                    "type": "string",
                    "example": "none"
                },
                "license_server": {
                    "type": "string",
                    "example": "licensing.openvpn.net:443"
                },
                "message": {
                    "type": "string",
                    "example": "Server info retrieved successfully"
                },
                "node_type": {
                    "type": "string",
                    "example": "PRIMARY"
                },
                "status": {
                    "type": "string",
                    "example": "healthy"
                },
                "web_server_name": {
                    "type": "string",
                    "example": "OpenVPN-AS"
                }
            }
        },
        "dto.SuccessResponse": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "object",
                    "properties": {
                        "data": {},
                        "message": {
                            "type": "string",
                            "example": "Operation successful"
                        },
                        "status": {
                            "type": "integer",
                            "example": 200
                        }
                    }
                }
            }
        },
...
        "dto.UserConnectionInfo": {
            "type": "object",
            "properties": {
                "connected_since": {
                    "type": "string",
                    "example": "2025-06-14T14:30:25Z"
                },
                "country": {
                    "type": "string",
                    "example": "Vietnam"
                },
                "real_address": {
                    "type": "string",
                    "example": "203.113.45.123"
                },
                "username": {
                    "type": "string",
                    "example": "testuser1"
                },
                "virtual_address": {
                    "type": "string",
                    "example": "172.27.232.15"
                }
            }
        },
...
        "dto.UserValidationError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "User is not currently connected"
                },
                "username": {
                    "type": "string",
                    "example": "testuser1"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and RSA256 JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}